se actualizan id a tipos long

+++ src/com/blo/sales/v2/controller/ICategoriesController.java
@@ -10,8 +10,8 @@ public interface ICategoriesController {
     
     WrapperIntPojoCategories getAllCategories() throws BloSalesV2Exception;
     
-    PojoIntCategory updateCategory(int id, PojoIntCategory newData) throws BloSalesV2Exception;
+    PojoIntCategory updateCategory(long id, PojoIntCategory newData) throws BloSalesV2Exception;
     
-    PojoIntCategory getCategoryById(int id) throws BloSalesV2Exception;
+    PojoIntCategory getCategoryById(long id) throws BloSalesV2Exception;
 
 }
